Change ObservedObject to StateObject in ContentView
Using StateObject ensures the view model will only be created once per view instance.

Also observe the scene phase from the Environment instead of notification center. We got lucky that the ObservedObject was initialized before the initial change of scenePhase from background to active, but the StateObject is created later, so we have to be sure to start loading even though we never see the value change initially.

Thanks for the tip @twostraws!
https://www.youtube.com/watch?v=qkcKTJhDyLs
https://developer.apple.com/documentation/swiftui/managing-model-data-in-your-app